Implement AivenObject for DynamicObject, eliminating need for specific types
All relevant objects have data in the same fields, and the DynamicObject allows generic access to those fields.
It makes much more sense to implement mutations on the basis of the generic access in DynamicObject, than
implementing the AivenObject trait for all Aiven types with identical code for each type.